// src/app/models/panel/school-year.ts
// Generated by https?://quicktype.io

export interface SchoolYearI {
  uid?:     string;
  schoolyearID?: string;
  schooltype?: string;
  schoolyear?: string;
  schoolyeartitle?: string;
  startingdate?: string;
  endingdate?: string;
  semestercode?: string;
  createDate?: string;
  modifyDate?: string;
  createUserID?: string;
  createUsername?: string;
  createUsertype?: string;
  dateCreated?: string;
}